+++ src/__init__.py
@@ -1,6 +1,6 @@
 """
 认证器盒子PC Tool
 """
-__version__ = "3.2.3"
+__version__ = "3.4.0"
 __author__ = "Autochips Inc"
 __description__ = "认证器盒子PC工具"